continue remove useless components after refactor
