--- crates/content-base/src/db/entity/relation.rs
@@ -1,29 +1,0 @@
-use serde::Deserialize;
-use surrealdb::sql::Thing;
-
-use crate::db::model::id::{ID, TB};
-
-#[derive(Debug, Deserialize, Clone, Hash, PartialEq, Eq)]
-pub struct RelationEntity {
-    id: Thing,
-    r#in: Thing,
-    out: Thing,
-}
-
-impl RelationEntity {
-    pub fn id(&self) -> String {
-        ID::from(self.id.clone()).id_with_table()
-    }
-    
-    pub fn in_table(&self) -> TB {
-        self.r#in.tb.as_str().into()
-    }
-
-    pub fn in_id(&self) -> String {
-        format!("{}:{}", self.r#in.tb, self.r#in.id.to_raw())
-    }
-
-    pub fn out_id(&self) -> String {
-        format!("{}:{}", self.out.tb, self.out.id.to_raw())
-    }
-}
