--- crates/content-base/src/db/entity/vector.rs
@@ -1,8 +1,0 @@
-use serde::Deserialize;
-use surrealdb::sql::Thing;
-
-#[derive(Debug, Deserialize)]
-pub(crate) struct VectorSearchEntity {
-    pub id: Thing,
-    pub distance: f32,
-}
